Moved AnkiBindingTagsDTO into the Application Layer.
